Refactor Move transactions code
